Attempt to fix ambiguity errors on Mac builds

+++ util/io/text_writer.cpp
@@ -69,42 +69,82 @@ TextWriter &operator<<(TextWriter &w, const bool v) {
 	return w;
 }
 
-TextWriter &operator<<(TextWriter &w, const int8_t v) {
+// TextWriter &operator<<(TextWriter &w, const int8_t v) {
+// 	w.write_i64(v);
+// 	return w;
+// }
+
+// TextWriter &operator<<(TextWriter &w, const int16_t v) {
+// 	w.write_i64(v);
+// 	return w;
+// }
+
+// TextWriter &operator<<(TextWriter &w, const int32_t v) {
+// 	w.write_i64(v);
+// 	return w;
+// }
+
+// TextWriter &operator<<(TextWriter &w, const int64_t v) {
+// 	w.write_i64(v);
+// 	return w;
+// }
+
+// TextWriter &operator<<(TextWriter &w, const uint8_t v) {
+// 	w.write_i64(v);
+// 	return w;
+// }
+
+// TextWriter &operator<<(TextWriter &w, const uint16_t v) {
+// 	w.write_i64(v);
+// 	return w;
+// }
+
+// TextWriter &operator<<(TextWriter &w, const uint32_t v) {
+// 	w.write_i64(v);
+// 	return w;
+// }
+
+// TextWriter &operator<<(TextWriter &w, const uint64_t v) {
+// 	w.write_i64(v);
+// 	return w;
+// }
+
+TextWriter &operator<<(TextWriter &w, const short v) {
 	w.write_i64(v);
 	return w;
 }
 
-TextWriter &operator<<(TextWriter &w, const int16_t v) {
+TextWriter &operator<<(TextWriter &w, const int v) {
 	w.write_i64(v);
 	return w;
 }
 
-TextWriter &operator<<(TextWriter &w, const int32_t v) {
+TextWriter &operator<<(TextWriter &w, const long v) {
 	w.write_i64(v);
 	return w;
 }
 
-TextWriter &operator<<(TextWriter &w, const int64_t v) {
+TextWriter &operator<<(TextWriter &w, const long long v) {
 	w.write_i64(v);
 	return w;
 }
 
-TextWriter &operator<<(TextWriter &w, const uint8_t v) {
+TextWriter &operator<<(TextWriter &w, const unsigned short v) {
 	w.write_i64(v);
 	return w;
 }
 
-TextWriter &operator<<(TextWriter &w, const uint16_t v) {
+TextWriter &operator<<(TextWriter &w, const unsigned int v) {
 	w.write_i64(v);
 	return w;
 }
 
-TextWriter &operator<<(TextWriter &w, const uint32_t v) {
+TextWriter &operator<<(TextWriter &w, const unsigned long v) {
 	w.write_i64(v);
 	return w;
 }
 
-TextWriter &operator<<(TextWriter &w, const uint64_t v) {
+TextWriter &operator<<(TextWriter &w, const unsigned long long v) {
 	w.write_i64(v);
 	return w;
 }
